Rename filter_files to find_files
And move to bluesearch.utils
