+++ tests/unit/test_utils.py
@@ -32,10 +32,40 @@ from bluesearch.utils import (
     JSONL,
     Timer,
     check_entity_type_consistency,
+    find_files,
     get_available_spacy_models,
     load_spacy_model,
 )
 
+
+class TestFindFiles:
+    def test_filtering_recursive(tmp_path):
+        input_path = pathlib.Path("tests/data/cord19_v35/document_parses/pdf_json/")
+        inputs = find_files(input_path, recursive=True)
+        filenames = sorted(x.name for x in inputs)
+        expected = [
+            "16e82ce0e0c8a1b36497afc0d4392b4fe21eb174.json",
+            "5f267fa1ef3a65e239aa974329e935a4d93dafd2.json",
+        ]
+        assert filenames == expected
+
+
+    def test_filtering(tmp_path):
+        input_path = pathlib.Path("tests/data/cord19_v35/")
+        inputs = find_files(input_path, recursive=True, match_filename=r"[a-z0-9]+\.json")
+        filenames = sorted(x.name for x in inputs)
+        expected = [
+            "16e82ce0e0c8a1b36497afc0d4392b4fe21eb174.json",
+            "5f267fa1ef3a65e239aa974329e935a4d93dafd2.json",
+        ]
+        assert filenames == expected
+
+
+    def test_filtering_empty(tmp_path):
+        message = "Value for argument 'match-filename' should not be empty!"
+        input_path = pathlib.Path("tests/data/cord19_v35/")
+        with pytest.raises(ValueError, match=message):
+            _ = find_files(input_path, recursive=True, match_filename="")
 
 class TestTimer:
     def test_errors(self):
